+++ src/app/demo/components-overview/table-server-example/table-server-example.component.ts
@@ -24,7 +24,11 @@ export class TableServerExampleComponent extends TableExampleBaseClass implement
   }
 
   onSelectedChange($event) {
-    console.log($event);
+    console.log('als Set:  ', $event);
+  }
+
+  onSelectedAsArrayChange($event) {
+    console.log('als Array:', $event);
   }
 
   reload() {