+++ src/app/modules/lux-common/lux-table/lux-table.component.ts
@@ -88,6 +88,7 @@ export class LuxTableComponent implements OnInit, AfterViewInit, DoCheck, OnDest
   @Input() luxPagerTooltip = '';
 
   @Output() luxSelectedChange: EventEmitter<Set<any>> = new EventEmitter<Set<any>>();
+  @Output() luxSelectedAsArrayChange: EventEmitter<any[]> = new EventEmitter<any[]>();
 
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
   @ViewChild(MatSort, { static: true }) sort: MatSort;
@@ -823,6 +824,7 @@ export class LuxTableComponent implements OnInit, AfterViewInit, DoCheck, OnDest
       this.lastSelectedEventData = newDataString;
 
       this.luxSelectedChange.next(this.luxSelected);
+      this.luxSelectedAsArrayChange.next(newData);
     }
   }