+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Bug Fixes
 - **lux-file-upload**: Upload-Komponente kann nicht im LUX-Dialog verwendet werden. [Issue 200](https://github.com/IHK-GfI/lux-components/issues/200)
+- **lux-lookup-label**: In blätterbaren Tabellen ändert sich der Wert nicht. [Issue 205](https://github.com/IHK-GfI/lux-components/issues/205)
 
 # Version 13.0.0
 ## New